Incorporate Searchbar class into my BlogGrid class and merge functionality
After 3 days of fighting the search bar is now rendering text results when the user searches based on blog_heading!

I didn't realise this was a time when Django required a specific naming convention for the function so had to rename it get_queryset.

Since I wanted to (maybe ambitiously) render the results of the user search on the main page rather than a search results page, I had to incorporate the logic into the BlogGrid class.

This now renders everything by default, then if a user searches something (picked up by search_query in the html form, and in the get_queryset method) it'll render just the blogs with headings related to that keyword. HURRAH!